Update util.js - Data Writes - Added autoskip over non-userdata section
These changes will automatically skip over the non-user data section of disc sectors making inserting large blocks of data less tedious. Might be an issue if someone tries to replace data in sectors that don't have the 280 or so bytes of error correction data for new FMV or maybe music but I wanted to note that here.